// client/src/interfaces/IUser.ts
export interface IUserLogin {
    account: string;
    password: string;
}

export interface IUserRegister extends IUserLogin {
    name: string;
    cf_password: string;
}

export interface IUser extends IUserLogin {
    avatar: string;
    createdAt: string;
    name: string;
    role: string;
    type: string;
    updatedAt: string;
    _id: string;
}